+++ suppaftp/src/sync_ftp/mod.rs
@@ -5,7 +5,7 @@
 mod data_stream;
 mod tls;
 
-use std::io::{copy, BufRead, BufReader, Cursor, Read, Write};
+use std::io::{BufRead, BufReader, Cursor, Read, Write, copy};
 #[cfg(not(feature = "secure"))]
 use std::marker::PhantomData;
 use std::net::{Ipv4Addr, SocketAddr, TcpListener, TcpStream, ToSocketAddrs};
@@ -23,9 +23,9 @@ pub use tls::{NativeTlsConnector, NativeTlsStream};
 #[cfg(feature = "rustls")]
 pub use tls::{RustlsConnector, RustlsStream};
 
+use super::Status;
 use super::regex::{EPSV_PORT_RE, MDTM_RE, PASV_PORT_RE, SIZE_RE};
 use super::types::{FileType, FtpError, FtpResult, Mode, Response};
-use super::Status;
 use crate::command::Command;
 #[cfg(feature = "secure")]
 use crate::command::ProtectionLevel;
@@ -1040,13 +1040,12 @@ mod test {
     #[cfg(feature = "secure")]
     use pretty_assertions::assert_eq;
     use rand::distr::Alphanumeric;
-    use rand::{rng, Rng};
-    use serial_test::serial;
+    use rand::{Rng, rng};
 
     use super::*;
+    use crate::FtpStream;
     use crate::test_container::SyncPureFtpRunner;
     use crate::types::FormatControl;
-    use crate::FtpStream;
 
     #[test]
     fn connect() {
@@ -1090,16 +1089,12 @@ mod test {
     }
 
     #[test]
-    #[serial]
     fn should_change_mode() {
-        crate::log_init();
-        let mut ftp_stream = FtpStream::connect("test.rebex.net:21")
-            .map(|x| x.active_mode(Duration::from_secs(30)))
-            .unwrap();
-        assert_eq!(ftp_stream.mode, Mode::Active);
-        assert_eq!(ftp_stream.active_timeout, Duration::from_secs(30));
-        ftp_stream.set_mode(Mode::Passive);
-        assert_eq!(ftp_stream.mode, Mode::Passive);
+        with_test_ftp_stream(|stream| {
+            assert_eq!(stream.mode, Mode::Passive);
+            stream.set_mode(Mode::Active);
+            assert_eq!(stream.mode, Mode::Active);
+        })
     }
 
     #[test]
@@ -1112,20 +1107,24 @@ mod test {
 
         let mut stream = FtpStream::connect_timeout(addr, Duration::from_secs(15)).unwrap();
         assert!(stream.login("test", "test").is_ok());
-        assert!(stream
-            .get_welcome_msg()
-            .unwrap()
-            .contains("220 You will be disconnected after 15 minutes of inactivity."));
+        assert!(
+            stream
+                .get_welcome_msg()
+                .unwrap()
+                .contains("220 You will be disconnected after 15 minutes of inactivity.")
+        );
     }
 
     #[test]
     fn welcome_message() {
         crate::log_init();
         with_test_ftp_stream(|stream| {
-            assert!(stream
-                .get_welcome_msg()
-                .unwrap()
-                .contains("220 You will be disconnected after 15 minutes of inactivity."));
+            assert!(
+                stream
+                    .get_welcome_msg()
+                    .unwrap()
+                    .contains("220 You will be disconnected after 15 minutes of inactivity.")
+            );
         });
     }
 
@@ -1141,10 +1140,12 @@ mod test {
     fn get_ref() {
         use std::time::Duration;
         with_test_ftp_stream(|stream| {
-            assert!(stream
-                .get_ref()
-                .set_read_timeout(Some(Duration::from_secs(10)))
-                .is_ok());
+            assert!(
+                stream
+                    .get_ref()
+                    .set_read_timeout(Some(Duration::from_secs(10)))
+                    .is_ok()
+            );
         });
     }
 
@@ -1196,9 +1197,11 @@ mod test {
     fn set_transfer_type() {
         with_test_ftp_stream(|stream| {
             assert!(stream.transfer_type(FileType::Binary).is_ok());
-            assert!(stream
-                .transfer_type(FileType::Ascii(FormatControl::Default))
-                .is_ok());
+            assert!(
+                stream
+                    .transfer_type(FileType::Ascii(FormatControl::Default))
+                    .is_ok()
+            );
         })
     }
 
@@ -1411,25 +1414,16 @@ mod test {
         format!("temp_{}", name)
     }
 
-    #[test]
-    fn test_should_set_passive_stream_builder() {
-        crate::log_init();
-        let _ftp_stream = FtpStream::connect("test.rebex.net:21")
-            .unwrap()
-            .passive_stream_builder(|addr| {
-                println!("Connecting to {}", addr);
-                TcpStream::connect(addr).map_err(FtpError::ConnectionError)
-            });
-    }
-
     /// Test if the stream is Send
     fn is_send<T: Send>(_send: T) {}
 
     fn is_sync<T: Sync>(_sync: T) {}
 
     #[test]
+    #[ignore = "just needs to compile"]
     fn test_ftp_stream_should_be_send() {
         crate::log_init();
+
         let ftp_stream = FtpStream::connect("test.rebex.net:21")
             .unwrap()
             .passive_stream_builder(|addr| {
@@ -1441,6 +1435,7 @@ mod test {
     }
 
     #[test]
+    #[ignore = "just needs to compile"]
     fn test_ftp_stream_should_be_sync() {
         crate::log_init();
         let ftp_stream = FtpStream::connect("test.rebex.net:21")